+++ core/src/test/scala/stasis/test/specs/unit/core/persistence/crates/CrateStoreSpec.scala
@@ -240,7 +240,7 @@ class CrateStoreSpec extends AsyncUnitSpec {
 
   private val testManifest = Manifest(
     crate = Crate.generateId(),
-    size = testContent.size,
+    size = testContent.size.toLong,
     copies = 4,
     source = Node.generateId(),
     origin = Node.generateId()
